requirements_py311_final.txt
absl-py==2.3.1
accelerate==0.34.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alabaster==1.0.0
alembic==1.17.0
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.11.0
asteroid-filterbanks==0.4.0
asttokens==3.0.1
attrs==25.4.0
audioread==3.0.1
av==15.1.0
babel==2.17.0
backcall==0.2.0
backports.functools-lru-cache==2.0.0
bangla==0.0.5
beautifulsoup4==4.14.3
black==25.9.0
bleach==6.3.0
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
cryptography==46.0.3
ctranslate2==4.6.0
customtkinter==5.2.2
cycler==0.12.1
cymem==2.0.11
Cython==3.1.6
darkdetect==0.8.0
dateparser==1.1.8
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
docopt==0.6.2
docutils==0.21.2
einops==0.8.1
encodec==0.1.1
executing==2.2.1
faster-whisper==1.2.0
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
g2pkk==0.1.2
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-genai==1.52.0
googleapis-common-protos==1.72.0
gpt4all==2.8.2
greenlet==3.2.4
grpcio==1.76.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.11
imagesize==1.4.1
img2pdf==0.6.3
indic-nlp-library==0.92
indic-transliteration==2.3.75
inflect==7.5.0
ipython==8.12.3
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.45.1
lxml==6.0.2
Mako==1.3.10
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
Morfessor==2.0.6
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.13
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==2.8.8
nltk==3.9.2
num2words==0.5.14
numba==0.62.1
oauthlib==3.3.1
ocrmypdf==16.12.0
omegaconf==2.3.0
onnxruntime==1.23.2
openai-whisper==20250625
optuna==4.5.0
packaging==25.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pdfminer.six==20251107
pefile==2023.2.7
pi_heif==1.1.1
pickleshare==0.7.5
pikepdf==10.0.2
pillow==12.0.0
pipreqs==0.5.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
preshed==3.0.10
primePy==1.3
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.1
pure_eval==0.2.3
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pynndescent==0.5.13
pyparsing==3.2.5
pypinyin==0.55.0
PyQt6==6.10.0
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.2
pyreadline3==3.5.4
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytokens==0.2.0
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.10.23
reindent==3.5.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
roman==5.1
roman-numerals-py==3.1.0
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
ruff==0.14.4
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.4.1
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.3
Sphinx==8.2.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.44
srsly==2.5.1
stack-data==0.6.3
SudachiDict-core==20250825
SudachiPy==0.6.10
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0+cu124
torchmetrics==1.8.2
torchvision==0.21.0+cu124
tornado==6.5.2
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
transformers==4.42.3
TTS==0.22.0
typeguard==4.4.4
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
whisperx==3.7.4
wrapt==2.0.0
yarg==0.1.9
yarl==1.22.0
yt-dlp==2025.10.22
